Move DiagnosticServerProxy classes into NETCore.Client library.
